Fix: #258 충돌해결 후 타입 불일치 에러 수정
- 타입 변경에 따른 초기값 수정
- nullable 제거 (Long? -> Long)

+++ app/src/main/java/com/konkuk/medicarecall/data/api/elders/MedicineService.kt
@@ -9,7 +9,7 @@ import de.jensklingenberg.ktorfit.http.Query
 interface MedicineService {
     @GET("elders/{elderId}/medication")
     suspend fun getDailyMedication(
-        @Path("elderId") elderId: Int,
+        @Path("elderId") elderId: Long,
         @Query("date") date: String,
     ): Response<MedicineResponseDto>
 }
