+++ app/src/main/java/com/konkuk/medicarecall/data/model/ElderIds.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ElderIds(
-    val elderIds: Map<Int, String>,
+    val elderIds: Map<Long, String>,
 )
