+++ app/src/main/java/com/konkuk/medicarecall/data/repository/HomeRepository.kt
@@ -3,6 +3,6 @@ package com.konkuk.medicarecall.data.repository
 import com.konkuk.medicarecall.domain.model.Home
 
 interface HomeRepository {
-    suspend fun requestImmediateCareCall(elderId: Int, careCallOption: String): Result<Unit>
-    suspend fun getHomeSummary(elderId: Int): Result<Home>
+    suspend fun requestImmediateCareCall(elderId: Long, careCallOption: String): Result<Unit>
+    suspend fun getHomeSummary(elderId: Long): Result<Home>
 }
