+++ app/src/main/java/com/konkuk/medicarecall/data/repository/MealRepository.kt
@@ -4,5 +4,5 @@ import com.konkuk.medicarecall.domain.model.Meal
 import java.time.LocalDate
 
 interface MealRepository {
-    suspend fun getMeals(elderId: Int, date: LocalDate): Result<List<Meal>>
+    suspend fun getMeals(elderId: Long, date: LocalDate): Result<List<Meal>>
 }
