+++ app/src/main/java/com/konkuk/medicarecall/data/repository/SetCallRepository.kt
@@ -4,6 +4,6 @@ import com.konkuk.medicarecall.data.dto.request.SetCallTimeRequestDto
 import com.konkuk.medicarecall.ui.model.CallTimes
 
 interface SetCallRepository {
-    suspend fun saveForElder(elderId: Int, body: SetCallTimeRequestDto): Result<Unit>
-    suspend fun saveForElder(elderId: Int, times: CallTimes): Result<Unit>
+    suspend fun saveForElder(elderId: Long, body: SetCallTimeRequestDto): Result<Unit>
+    suspend fun saveForElder(elderId: Long, times: CallTimes): Result<Unit>
 }
