+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/statistics/viewmodel/StatisticsViewModel.kt
@@ -70,7 +70,7 @@ class StatisticsViewModel(
         }
 
         val id = _uiState.value.selectedElderId
-        if (id == -1) return
+        if (id == -1L) return
         val start = _uiState.value.currentWeek.first
         getWeeklyStatistics(
             elderId = id,
@@ -79,7 +79,7 @@ class StatisticsViewModel(
         )
     }
 
-    fun setSelectedElderId(id: Int) {
+    fun setSelectedElderId(id: Long) {
         if (_uiState.value.selectedElderId != id) {
             // [수정 3] 새로운 사용자를 선택하면 earliestDate를 초기화합니다.
             // 이렇게 해야 이전 사용자의 기록이 다음 사용자에게 영향을 주지 않습니다.
@@ -108,7 +108,7 @@ class StatisticsViewModel(
     }
 
     private fun getWeeklyStatistics(
-        elderId: Int,
+        elderId: Long,
         startDate: LocalDate,
         ignoreLoadingGate: Boolean = false,
     ) {
